% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csestudy-package.R
\docType{package}
\name{csestudy-package}
\alias{csestudy-package}
\title{csestudy: Time-Series Inference for Cross-Sectional Event Studies}
\description{
R implementation of the methodology of Cohn, Johnson, Liu and Wardlaw
(2026), "Past is Prologue: Inference from the Cross Section of Returns
Around an Event," \emph{Journal of Financial Economics} 180, 104278
\doi{10.1016/j.jfineco.2026.104278}.
}
\details{
The package mirrors the Stata (\code{csestudy.ado} / \code{csestudy.mata}) and Python
(\code{CSEventStudy}) implementations shipped in the same repository and reproduces
their numerical output on the shared sample data.
}
\seealso{
\code{\link[=csestudy]{csestudy()}}
}
\author{
\strong{Maintainer}: Malcolm I. Wardlaw \email{malcolm.wardlaw@uga.edu}

Authors:
\itemize{
  \item Malcolm I. Wardlaw \email{malcolm.wardlaw@uga.edu}
  \item Jonathan B. Cohn
  \item Travis L. Johnson
  \item Zack Liu
}

}
\keyword{internal}
